# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PurrfectPets",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.zooshop.local:8443/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Получение заказов пользователя",
                "description": "Получение заказов пользователя",
                "operationId": "479e95fb191e4a4e1a912114996874bf",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ успешно создан."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/order/create": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Создание заказа",
                "description": "Создание заказа",
                "operationId": "c1e40c5476093f70df5112136f2ddd32",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "orderProperties": {
                                                "properties": {
                                                    "FIO": {
                                                        "type": "string",
                                                        "example": "Морозов Николай Михайлович"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "morozov@test.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+7(999)999-99-99"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Москва, ул. Ленина, д. 1"
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "example": "Комментарий"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "products": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/components/schemas/OrderProduct"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "properties": {
                                                "orderId": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ успешно создан."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/order/{orderId}/cancel": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Отмена заказа",
                "description": "Отмена заказа",
                "operationId": "9793519b7da98335aa8605cf49949589",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ успешно отменён."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не существует",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Menu type model",
                "required": [
                    "id",
                    "user_id",
                    "status",
                    "is_paid",
                    "price"
                ],
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int128",
                        "example": 1
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "User ID",
                        "type": "integer",
                        "format": "int128",
                        "example": 1
                    },
                    "status": {
                        "title": "Status",
                        "description": "Status for order",
                        "type": "integer",
                        "example": 1
                    },
                    "is_paid": {
                        "title": "Active",
                        "description": "$is_paid",
                        "type": "boolean",
                        "example": true
                    },
                    "price": {
                        "title": "Price",
                        "description": "Price",
                        "type": "number",
                        "format": "float",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-04-08 00:00:00"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-04-08 00:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Order"
                }
            },
            "OrderProduct": {
                "title": "OrderProduct",
                "required": [
                    "product_id",
                    "count"
                ],
                "properties": {
                    "product_id": {
                        "title": "Product ID",
                        "description": "$product_id",
                        "type": "integer",
                        "format": "int128",
                        "example": 1
                    },
                    "count": {
                        "title": "Count",
                        "description": "$count",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderProduct"
                }
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}